requirements.txt
fastapi
uvicorn[standard]

sqlalchemy>=2.0
asyncpg

python-jose[cryptography]

pydantic-settings
email-validator

alembic

pytest
pytest-asyncio
httpx

passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

